feat: adicionar testes

+++ database/migrations/2026_03_16_123702_remove_completed_from_tasks_table.php
@@ -7,9 +7,11 @@ use Illuminate\Support\Facades\Schema;
 return new class () extends Migration {
     public function up(): void
     {
-        Schema::table('tasks', function (Blueprint $table) {
-            $table->dropColumn('completed');
-        });
+        if (Schema::hasColumn('tasks', 'completed')) {
+            Schema::table('tasks', function (Blueprint $table) {
+                $table->dropColumn('completed');
+            });
+        }
     }
 
     public function down(): void
